requirements.txt
docker
python-dotenv
autogen-core>=0.4.0dev1
autogen-agentchat>=0.4.0dev1
autogen-ext>=0.4.0dev1
json_repair
tenacity
tavily-python
firecrawl-py
ultralytics 
opencv-python 
matplotlib
pdf2image
numpy
azure-ai-vision-face
rich
azure-cosmos
azure-mgmt-cosmosdb
azure-core
azure-identity
uvicorn
fastapi
python-multipart
streamlit
pillow
azure-storage-blob
requests